refactor, add skeleton for IT
